+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dispatch queue list and Scheduled list on Dispatch tab
 - Virtual list
 - `Referendums list`, `External list`, `Proposals list` on `Democracy page`
+- Progress info block
 
 #### Fixed
 - Collected winning data on auction tab